+++ CHANGES.rst
@@ -72,6 +72,9 @@ mast
 - Added ``return_uri_map`` parameter to ``Observations.get_cloud_uris`` to return a mapping of the input data product URIs 
   to the returned cloud URIs. [#3314]
 
+- Added ``verbose`` parameter to ``Observations.get_cloud_uris`` to control whether warnings are logged when a product cannot 
+  be found in the cloud. [#3314]
+
 
 Infrastructure, Utility and Other Changes and Additions
 -------------------------------------------------------